pyproject.toml
[tool.poetry]
name = "foodie-map"
version = "0.1.0"
description = "A map where you can save your favorite restaurants as bookmarks"
authors = ["Tsy-Jon Leslie Lau <tsyjonlau@gmail.com>"]
license = "MIT"
readme = "./README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
django-cors-headers = "^4.2.0"
djangorestframework = "^3.14.0"
whitenoise = "^6.5.0"
psycopg = "^3.1.11"
django-webpack-loader = "^2.0.1"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
